server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
marshmallow==3.20.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
click==8.1.7
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4 